# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: НА_УТРАХ_4 Watch Configurator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /configurations:
    post:
      operationId: createConfiguration
      tags: [configurations]
      summary: Create a watch configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchConfigInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchConfig"
  /configurations/{id}:
    get:
      operationId: getConfiguration
      tags: [configurations]
      summary: Get a configuration by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchConfig"
        "404":
          description: Not found
    put:
      operationId: updateConfiguration
      tags: [configurations]
      summary: Update a configuration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchConfigUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchConfig"
...
  /analytics/time-series:
    get:
      operationId: getAnalyticsTimeSeries
      tags: [analytics]
      summary: Get time-series analytics data for graphs
      parameters:
        - name: metric
          in: query
          schema:
            type: string
            enum: [revenue, orders, visitors, conversions]
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Time series data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeSeriesPoint"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login as admin or courier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not authenticated
  /couriers:
    get:
      operationId: listCouriers
      tags: [auth]
      summary: List all couriers (admin only)
      responses:
        "200":
          description: Couriers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
    post:
      operationId: createCourier
      tags: [auth]
      summary: Create a courier (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourierInput"
      responses:
        "201":
          description: Courier created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
components:
  schemas:
...
    WatchConfig:
      type: object
      required: [id, sessionId, watchfaceGeometry, watchfaceMaterial, watchfaceColor, braceletMaterial, braceletType, braceletColor, handsEnabled, createdAt]
      properties:
        id:
          type: integer
        sessionId:
          type: string
        presetId:
          type: ["integer", "null"]
        watchfaceGeometry:
          type: string
          enum: [circle, square, star, drawn]
        watchfaceMaterial:
          type: string
          enum: [plastic, metal]
        watchfaceColor:
          type: string
        watchfaceSize:
          type: ["string", "null"]
        braceletMaterial:
          type: string
          enum: [plastic_solid, plastic_segmented, metal_solid, metal_segmented, resin, leather, cotton_fabric]
        braceletType:
          type: string
          enum: [solid, segmented]
        braceletColor:
          type: string
        braceletWidth:
          type: ["string", "null"]
        braceletLength:
          type: ["string", "null"]
        closingMechanism:
          type: ["string", "null"]
        handsStyle:
          type: ["string", "null"]
        handsColor:
          type: ["string", "null"]
        handsEnabled:
          type: boolean
        serialNumber:
          type: ["string", "null"]
        customWatchfaceUrl:
          type: ["string", "null"]
        customBackgroundUrl:
          type: ["string", "null"]
        skinFullUrl:
          type: ["string", "null"]
        skinStripeUrl:
          type: ["string", "null"]
        boxType:
          type: ["string", "null"]
        boxColor:
          type: ["string", "null"]
        createdAt:
          type: string
    WatchConfigInput:
      type: object
      required: [sessionId, watchfaceGeometry, watchfaceMaterial, watchfaceColor, braceletMaterial, braceletType, braceletColor, handsEnabled]
      properties:
        sessionId:
          type: string
        presetId:
          type: ["integer", "null"]
        watchfaceGeometry:
          type: string
          enum: [circle, square, star, drawn]
        watchfaceMaterial:
          type: string
          enum: [plastic, metal]
        watchfaceColor:
          type: string
        watchfaceSize:
          type: string
        braceletMaterial:
          type: string
          enum: [plastic_solid, plastic_segmented, metal_solid, metal_segmented, resin, leather, cotton_fabric]
        braceletType:
          type: string
          enum: [solid, segmented]
        braceletColor:
          type: string
        braceletWidth:
          type: string
        braceletLength:
          type: string
        closingMechanism:
          type: string
        handsStyle:
          type: string
        handsColor:
          type: string
        handsEnabled:
          type: boolean
        serialNumber:
          type: string
        customWatchfaceUrl:
          type: string
        customBackgroundUrl:
          type: string
        skinFullUrl:
          type: string
        skinStripeUrl:
          type: string
        boxType:
          type: string
        boxColor:
          type: string
    WatchConfigUpdate:
      type: object
      properties:
        watchfaceGeometry:
          type: string
          enum: [circle, square, star, drawn]
        watchfaceMaterial:
          type: string
          enum: [plastic, metal]
        watchfaceColor:
          type: string
        watchfaceSize:
          type: string
        braceletMaterial:
          type: string
          enum: [plastic_solid, plastic_segmented, metal_solid, metal_segmented, resin, leather, cotton_fabric]
        braceletType:
          type: string
          enum: [solid, segmented]
        braceletColor:
          type: string
        braceletWidth:
          type: string
        braceletLength:
          type: string
        closingMechanism:
          type: string
        handsStyle:
          type: string
        handsColor:
          type: string
        handsEnabled:
          type: boolean
        serialNumber:
          type: string
        customWatchfaceUrl:
          type: string
        customBackgroundUrl:
          type: string
        skinFullUrl:
          type: string
        skinStripeUrl:
          type: string
        boxType:
          type: string
        boxColor:
          type: string
...
    TimeSeriesPoint:
      type: object
      required: [date, value]
      properties:
        date:
          type: string
        value:
          type: number
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    LoginResult:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/AdminUser"
    AdminUser:
      type: object
      required: [id, username, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [admin, courier]
        createdAt:
          type: string
    CourierInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string