# lib/api-zod/src/generated/types/watchConfigUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * НА_УТРАХ_4 Watch Configurator API
 * OpenAPI spec version: 0.1.0
 */
import type { WatchConfigUpdateBraceletMaterial } from './watchConfigUpdateBraceletMaterial';
import type { WatchConfigUpdateBraceletType } from './watchConfigUpdateBraceletType';
import type { WatchConfigUpdateWatchfaceGeometry } from './watchConfigUpdateWatchfaceGeometry';
import type { WatchConfigUpdateWatchfaceMaterial } from './watchConfigUpdateWatchfaceMaterial';

export interface WatchConfigUpdate {
  watchfaceGeometry?: WatchConfigUpdateWatchfaceGeometry;
  watchfaceMaterial?: WatchConfigUpdateWatchfaceMaterial;
  watchfaceColor?: string;
  watchfaceSize?: string;
  braceletMaterial?: WatchConfigUpdateBraceletMaterial;
  braceletType?: WatchConfigUpdateBraceletType;
  braceletColor?: string;
  braceletWidth?: string;
  braceletLength?: string;
  closingMechanism?: string;
  handsStyle?: string;
  handsColor?: string;
  handsEnabled?: boolean;
  serialNumber?: string;
  customWatchfaceUrl?: string;
  customBackgroundUrl?: string;
  skinFullUrl?: string;
  skinStripeUrl?: string;
  boxType?: string;
  boxColor?: string;
}